[devmgr] Relocate all mount-related functionality to vfs-mount
- Functions are unchanged, just relocated. More {un}mount-related work
  to come.
